frontend: add a form in the dialog
delete the secondFactorLogin page
